fix: Install `orgmode` plugin to provide `org` treesitter parser during testing
The `org` TS parser was dropped from `nvim-treesitter` due to it being
unmaintained. However, the popular `orgmode` plugin maintains a fork
that it installs. This installs the plugin in the test harness script to
allow testing of the org queries.

Also converts the expected test output to the new format.